feat: add segmental channels
